requirements.txt
requests>=2.32.0
beautifulsoup4>=4.14.0
python-dotenv>=1.2.0